feat: independent token strategy for experimental condition encoding (P1-P3)
Cond-token pipeline complete: backward compat + training + inference API.

P1: Backward compatibility & integration
- Verified ssl_model.ckpt loads into cond-token model: all 65 backbone
  weights identical, 5 cond-token params freshly initialized
- Forward pass: 61 peaks -> 63 tokens with +2 cond tokens
- SSL step produces finite loss
- Bug fix: nn.Embedding requires LongTensor but HDF5 stores adduct
  as np.int16; added .long() cast in dreams.py:202

P2: Training verification
- End-to-end pipeline verified: data -> forward -> loss -> backward -> grad -> optimizer
- Gradients flow to both adduct_embedding and ce_mlp (all ACTIVE)
- No NaN/Inf losses across 5 batches

P3: Inference API upgrade
- MGF parser extracts COLLISION_ENERGY + CE fields (io.py)
- RawSpectraDataset supports adduct+CE; MSData.to_torch_dataset()
  auto-detects and passes column data (data.py)
- _load_dreams_checkpoint() bypasses PyTorch 2.6 weights_only issue (api.py)
- from_pretrained_backbone() loads backbone with optional cond tokens (api.py)
- dreams_embeddings() accepts enable_cond_tokens, adduct_vocab_size, ce_max (api.py)

+++ dreams/models/dreams/dreams.py
@@ -199,7 +199,7 @@ class DreaMS(pl.LightningModule):
             if adduct is None or collision_energy is None:
                 raise ValueError('adduct and collision_energy required when enable_cond_tokens=True')
             # Adduct embedding: (batch,) -> (batch, 1, d_model)
-            adduct_embs = self.adduct_embedding(adduct).unsqueeze(1)
+            adduct_embs = self.adduct_embedding(adduct.long()).unsqueeze(1)
             # CE MLP: (batch,) -> (batch, 1, 1) -> (batch, 1, d_model)
             ce_embs = self.ce_mlp(collision_energy.unsqueeze(-1).unsqueeze(-1))
             # Build new sequence: [precursor(0), adduct(1), CE(2), fragment_peaks(3...)]
